refactor: add EventSort utils class.
