+++ src/const.js
@@ -23,38 +23,8 @@ const FilterType = {
   PAST: 'past'
 };
 
-const sortTypes = [
-  {
-    name: 'day',
-    isAvailable: true,
-    isDefault: true
-  },
-  {
-    name: 'event',
-    isAvailable: false,
-    isDefault: false
-  },
-  {
-    name: 'time',
-    isAvailable: true,
-    isDefault: false
-  },
-  {
-    name: 'price',
-    isAvailable: true,
-    isDefault: false
-  },
-  {
-    name: 'offers',
-    isAvailable: false,
-    isDefault: false
-  }
-];
-
-
 export {
   EVENT_TYPES,
   DateFormat,
   FilterType,
-  sortTypes
 };
