+++ src/view/sort-view.js
@@ -1,4 +1,3 @@
-import {sortTypes} from '../const';
 import {getCapitalizedString} from '../utils/common-utils.js';
 import AbstractView from '../framework/view/abstract-view.js';
 
@@ -23,7 +22,7 @@ function createSortItemTemplate(sortItem) {
   );
 }
 
-function createSortTemplate() {
+function createSortTemplate(sortTypes) {
   const sortItems = sortTypes.slice().map((sortItem) => createSortItemTemplate(sortItem)).join('');
   return (
     `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
@@ -34,17 +33,19 @@ function createSortTemplate() {
 
 
 export default class SortView extends AbstractView {
+  #sortTypes = null;
   #onSortTypeChangeClick = null;
 
-  constructor({onSortTypeChangeClick}) {
+  constructor({sortTypes, onSortTypeChangeClick}) {
     super();
+    this.#sortTypes = sortTypes;
     this.#onSortTypeChangeClick = onSortTypeChangeClick;
 
     this.element.addEventListener('click', this.#sortTypeChangeHandler);
   }
 
   get template() {
-    return createSortTemplate();
+    return createSortTemplate(this.#sortTypes);
   }
 
   #sortTypeChangeHandler = (evt) => {
